requirements.txt
python-telegram-bot
google-api-python-client
requests
openai
python-dotenv